Fix variable in barcucci.sabina

+++ barcucci.sabina/final_project_folder/code/pod_04/MQ7.ino
@@ -1,4 +1,4 @@
-float leggigas () {
+int leggigas () {
   gas = analogRead (MQ7);
   Serial.println (gas);
   return gas;
